refactor(test): 패키지 이동에 따른 테스트 코드 위치 및 import 수정
