+++ src/test/java/com/dnd/moddo/domain/expense/entity/ExpenseTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import com.dnd.moddo.domain.group.entity.Group;
-import com.dnd.moddo.support.GroupTestFactory;
+import com.dnd.moddo.global.support.GroupTestFactory;
 
 class ExpenseTest {
