+++ src/test/java/com/dnd/moddo/domain/expense/service/implementation/ExpenseCreatorTest.java
@@ -20,7 +20,7 @@ import com.dnd.moddo.domain.expense.repository.ExpenseRepository;
 import com.dnd.moddo.domain.group.entity.Group;
 import com.dnd.moddo.domain.group.repository.GroupRepository;
 import com.dnd.moddo.domain.memberExpense.service.implementation.MemberExpenseValidator;
-import com.dnd.moddo.support.GroupTestFactory;
+import com.dnd.moddo.global.support.GroupTestFactory;
 
 @ExtendWith(MockitoExtension.class)
 class ExpenseCreatorTest {
