+++ src/test/java/com/dnd/moddo/domain/groupMember/entity/GroupMemberTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 
 import com.dnd.moddo.domain.group.entity.Group;
 import com.dnd.moddo.domain.groupMember.entity.type.ExpenseRole;
-import com.dnd.moddo.support.GroupTestFactory;
+import com.dnd.moddo.global.support.GroupTestFactory;
 
 class GroupMemberTest {
